replace old main with the new main file
